refactor: 폴더 rename
